requirements.txt
Flask==1.0.2
uWSGI==2.0.17
RPi.GPIO==0.6.5
